Task Component

See merge request Time-Capture-System/TCS-MobileApp!6

+++ lib/models/task/task.dart
@@ -9,7 +9,7 @@ class Task {
 
   String taskName;
 
-  int estimatedHours;
+  double estimatedHours;
 
   DateTime createdAt;
